api/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
pydantic>=2.10
pydantic-settings>=2.7
python-multipart>=0.0.20
httpx>=0.28
sqlalchemy>=2.0
psycopg[binary]>=3.2.10
pgvector>=0.3
google-genai>=0.3
openai>=1.59
pymupdf>=1.25
readability-lxml>=0.8.1
lxml-html-clean>=0.4
markdown>=3.7
python-docx>=1.1.2
weasyprint>=63
pypandoc>=1.14
pyyaml>=6
tenacity>=9
